moved vote out of dill section and added vot to sidebar navigation
