+++ v2/components/NavItems.tsx
@@ -6,7 +6,8 @@ import {
   SpeakerphoneIcon,
   TemplateIcon,
   UserGroupIcon,
-  HeartIcon
+  StarIcon,
+  HeartIcon,
 } from "@heroicons/react/solid";
 import { useTranslation } from "next-i18next";
 
@@ -27,6 +28,7 @@ const NavItems: FC<Props> = ({ onClick }) => {
       icon: LightningBoltIcon,
     },
     { name: t("v2.nav.dill"), href: "/v2/dill", icon: UserGroupIcon },
+    { name: t("v2.nav.vote"), href: "/v2/vote", icon: StarIcon },
     { name: t("v2.nav.brinery"), href: "/v2/brinery", icon: HeartIcon },
     { name: t("v2.nav.stats"), href: "/v2/stats", icon: ChartBarIcon },
     { name: t("v2.nav.faq"), href: "#", icon: QuestionMarkCircleIcon },
