+++ v2/features/vote/CastVoteMainnet.tsx
@@ -2,8 +2,7 @@ import { Web3Provider } from "@ethersproject/providers";
 import { Contract, BigNumber } from "ethers";
 import { toast, ToastOptions } from "react-toastify";
 import { PickleModelJson } from "picklefinance-core";
-import gaugeProxyAbi from "../../../../containers/ABIs/gauge-proxy.json";
-import { JarDefinition } from "picklefinance-core/lib/model/PickleModelJson";
+import gaugeProxyAbi from "../../../containers/ABIs/gauge-proxy.json";
 import { findJar } from "v2/store/core.helpers";
 
 const GAUGE_PROXY = "0x2e57627ACf6c1812F99e274d0ac61B786c19E74f";
