+++ v2/features/vote/ChainTableRow.tsx
