+++ v2/features/vote/ChainWeightPieChart.tsx
