+++ v2/features/vote/ChainWeightTable.tsx
