+++ v2/features/vote/JarTableRow.tsx
