+++ v2/features/vote/StratTable.tsx
