+++ v2/features/vote/StratTableHeader.tsx
