+++ v2/features/vote/StratTableRow.tsx
